@@ -9,32 +9,31 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "json-server": "json-server db.json",
-    "fix-db": "rm db.json; cp db.json.fixture ./db.json"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/forms": "^4.4.6",
-    "@angular/http": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@angular/platform-browser-dynamic": "^4.4.6",
-    "@angular/router": "^4.4.6",
+    "@angular/common": "^5.2.4",
+    "@angular/compiler": "^5.2.4",
+    "@angular/core": "^5.2.4",
+    "@angular/forms": "^5.2.4",
+    "@angular/http": "^5.2.4",
+    "@angular/platform-browser": "^5.2.4",
+    "@angular/platform-browser-dynamic": "^5.2.4",
+    "@angular/router": "^5.2.4",
     "core-js": "^2.5.3",
     "enhanced-resolve": "^3.4.1",
     "json-server": "^0.11.2",
     "rxjs": "^5.5.6",
+    "typeorm": "^0.2.0-alpha.18",
     "zone.js": "^0.8.20"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.6",
-    "@angular/compiler-cli": "^4.4.6",
+    "@angular/cli": "^1.6.8",
+    "@angular/compiler-cli": "^5.2.4",
     "@types/jasmine": "2.5.38",
-    "@types/node": "^6.0.96",
-    "codelyzer": "~2.0.0",
+    "@types/node": "^6.0.100",
+    "codelyzer": "~4.0.0",
     "concurrently": "^3.5.1",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "~3.2.0",
@@ -42,12 +41,14 @@
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^0.2.0",
+    "karma-hue-reporter": "0.0.4",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "node-sass": "^4.7.2",
     "protractor": "~5.1.0",
+    "ts-mockito": "^2.2.9",
     "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "^2.3.4"
+    "tslint": "~5.7.0",
+    "typescript": "2.6.2"
   }
 }
