@@ -10,8 +10,15 @@ module.exports = function (config) {
       require('karma-chrome-launcher'),
       require('karma-jasmine-html-reporter'),
       require('karma-coverage-istanbul-reporter'),
+      require('karma-hue-reporter'),
       require('@angular/cli/plugins/karma')
     ],
+    customLaunchers: {
+      Chrome_with_files: {
+        base: 'Chrome',
+        flags: ['--allow-file-access-from-files']
+      }
+    },
     client: {
       clearContext: false // leave Jasmine Spec Runner output visible in browser
     },
@@ -28,17 +35,24 @@ module.exports = function (config) {
       reports: ['html', 'lcovonly'],
       fixWebpackSourcePaths: true
     },
+    webpack: { node: { fs: 'empty' } },
     angularCli: {
       environment: 'dev'
     },
+    hueReporter: {
+      ip: '192.168.1.10',
+      user: 'Ba8MqTZlNf8ywuWIfnqKxEsxfdBxhrN8eVoJES6h',
+      applyTo: 'groups',
+      applyToId: 4
+    },
     reporters: config.angularCli && config.angularCli.codeCoverage
-      ? ['progress', 'coverage-istanbul']
-      : ['progress', 'kjhtml'],
+      ? ['progress', 'coverage-istanbul', 'hue']
+      : ['progress', 'kjhtml', 'hue'],
     port: 9876,
     colors: true,
     logLevel: config.LOG_INFO,
     autoWatch: true,
-    browsers: ['Chrome'],
+    browsers: ['Chrome_with_files'],
     singleRun: false
   });
 };
