@@ -1,4 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup } from '@angular/forms';
+import { FormConfig, FormsSchemaService } from '../forms-schema.service';
 
 @Component({
   selector: 'app-base-form',
@@ -6,10 +8,20 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./base-form.component.scss']
 })
 export class BaseFormComponent implements OnInit {
+  config: FormConfig;
+  form: FormGroup;
 
-  constructor() { }
+  constructor(
+    private fsc: FormsSchemaService
+  ) { }
 
   ngOnInit() {
+    this.fsc.getConfig().subscribe(config => {
+      this.fsc.getFg().subscribe(fg => {
+        this.config = config;
+        this.form = fg;
+      });
+    });
   }
 
 }
