@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { Sequence } from '../../../models/sequence';
+import { Sequence } from '../../../models/Sequence';
 
 @Component({
   selector: 'app-sequence-thumbnail',
